aider: Call `summarizeAbstracts()` after loading JSON files in `loadJsonFiles()` function.
# Aider chat conversation:

USER: So the load json files button should load all the abstracts in that div "summarize-results".

ASSISTANT: To make the "Load JSON Files" button load all the abstracts into the "summarize-results" div, we need to modify the `loadJsonFiles()` function in `static/scripts.js` to call the `summarizeAbstracts()` function after loading the JSON files.

Here is the *SEARCH/REPLACE* block for that:

static/scripts.js
```javascript
<<<<<<< SEARCH
    .catch(e => {
        console.log('There was a problem with the fetch operation: ' + e.message);
    });
=======
    .then(() => {
        summarizeAbstracts();
    })
    .catch(e => {
        console.log('There was a problem with the fetch operation: ' + e.message);
    });
>>>>>>> REPLACE
```

This change will cause the `summarizeAbstracts()` function to be called immediately after the JSON files are loaded, which will load the abstracts into the "summarize-results" div.