aider: Refactored `loadJsonFiles()` to only load the contents of the selected JSON file into the "search-results" div.
# Aider chat conversation:

USER: Ok, so we have a problem. The contents of the dropdown are showing up inside "search-results". We only want to load the contents of the selected json files in the "search-results" div.

ASSISTANT: To fix this issue, we need to modify the `loadJsonFiles()` function in `static/scripts.js` to only load the contents of the selected JSON file into the "search-results" div, not the names of all JSON files.

Here is the *SEARCH/REPLACE* block for that:

static/scripts.js
```javascript
<<<<<<< SEARCH
function loadJsonFiles() {
    console.log('loadJsonFiles called');
    fetch('/load_json_files')
    .then(response => {
        if (!response.ok) {
            throw new Error(`HTTP error! status: ${response.status}`);
        }
        return response.json();
    })
    .then(data => {
        const resultsElement = document.getElementById('search-results');
        resultsElement.innerHTML = '';
        data.files.forEach(file => {
            const p = document.createElement('p');
            p.textContent = file;
            resultsElement.appendChild(p);
            const br = document.createElement('br');
            resultsElement.appendChild(br);
        });
    })
    .then(() => {
        summarizeAbstracts();
    })
    .catch(e => {
        console.log('There was a problem with the fetch operation: ' + e.message);
    });
}
=======
function loadJsonFiles() {
    console.log('loadJsonFiles called');
    const filePath = document.getElementById('json-files').value;
    fetch('/load_json_files', {
        method: 'POST',
        headers: {
            'Content-Type': 'application/json'
        },
        body: JSON.stringify({ json_file_path: filePath })
    })
    .then(response => {
        if (!response.ok) {
            throw new Error(`HTTP error! status: ${response.status}`);
        }
        return response.json();
    })
    .then(data => {
        const resultsElement = document.getElementById('search-results');
        resultsElement.innerHTML = JSON.stringify(data, null, 2);
    })
    .catch(e => {
        console.log('There was a problem with the fetch operation: ' + e.message);
    });
}
>>>>>>> REPLACE
```

This change will make the `loadJsonFiles()` function send a POST request to the `/load_json_files` endpoint with the selected JSON file path. The server should then return the contents of the selected JSON file, which will be displayed in the "search-results" div.