aider: Populate dropdown with JSON file names when the page loads.
# Aider chat conversation:

USER: I am not able to see any of the dropdown items.

ASSISTANT: To fix the issue with the dropdown not showing any items, we need to modify the `loadJsonFiles()` function in `static/scripts.js` to populate the dropdown with the names of the JSON files when the page loads.

Here is the *SEARCH/REPLACE* block for that:

static/scripts.js
```python
<<<<<<< SEARCH
window.onload = function() {
    loadJsonFiles();
}
=======
window.onload = function() {
    fetch('/load_json_files')
    .then(response => response.json())
    .then(data => {
        const dropdown = document.getElementById('json-files');
        data.files.forEach(file => {
            const option = document.createElement('option');
            option.text = file;
            dropdown.add(option);
        });
    });
}
>>>>>>> REPLACE
```

This change will make the `window.onload` function send a GET request to the `/load_json_files` endpoint when the page loads. The server should then return the names of all JSON files, which will be added to the dropdown.